adds season stats mod and test
