+++ spec/stat_tracker_spec.rb
@@ -29,125 +29,8 @@ RSpec.describe StatTracker do
     end
   end
 
-  
 
-  context 'league stats methods' do
-    game_path = './spec/fixture_files/test_games.csv'
-    team_path = './spec/fixture_files/test_teams.csv'
-    game_teams_path = './spec/fixture_files/test_game_teams.csv'
-    locations = {
-      games: game_path,
-      teams: team_path,
-      game_teams: game_teams_path
-    }
-    stat_tracker = StatTracker.from_csv(locations)
 
-    it '#count_of_teams' do
-      expect(stat_tracker.count_of_teams).to eq(32)
-    end
-
-    it '#games_by_team_id' do
-      expect(stat_tracker.games_by_team_id).to be_a Hash
-    end
-
-#mock and stub
-    it '#goals_per_team' do
-      expect(stat_tracker.goals_per_team["3"]).to eq([2, 2, 1, 2, 1, 1, 0, 2, 2, 1, 1, 3, 2, 2, 1, 0, 3, 2, 3, 2, 2, 2, 2, 2, 3, 3, 0, 5, 0, 1, 3, 0, 1, 2, 2, 2, 3, 2, 1, 1, 3, 0])
-    end
-
-    it '#team_name_by_team_id' do
-      expect(stat_tracker.team_name_by_team_id("1")).to eq("Atlanta United")
-    end
-
-    it '#best_offense' do
-      expect(stat_tracker.best_offense).to eq("FC Dallas")
-    end
-
-    it '#worst_offense' do
-      expect(stat_tracker.worst_offense).to eq("Sky Blue FC")
-    end
-
-# MOCK AND STUB PLZ
-    it '#finds_home_games' do
-      expect(stat_tracker.games_by_hoa("home")["3"]).to eq([20, 32])
-    end
-
-    it '#highest_scoring_visitor' do
-      expect(stat_tracker.highest_scoring_visitor).to eq("Columbus Crew SC")
-    end
-
-    it '#highest_scoring_home_team' do
-      expect(stat_tracker.highest_scoring_home_team).to eq("San Jose Earthquakes")
-    end
-
-    it '#lowest_scoring_visitor' do
-      expect(stat_tracker.lowest_scoring_visitor).to eq("Chicago Fire")
-    end
-
-    it '#lowest_scoring_home_team' do
-      expect(stat_tracker.lowest_scoring_home_team).to eq("Washington Spirit FC")
-    end
-  end
-
-  context 'Season stats methods' do
-    game_path = './spec/fixture_files/test_games.csv'
-    team_path = './spec/fixture_files/test_teams.csv'
-    game_teams_path = './spec/fixture_files/test_game_teams.csv'
-    locations = {
-      games: game_path,
-      teams: team_path,
-      game_teams: game_teams_path
-    }
-    stat_tracker = StatTracker.from_csv(locations)
-
-    it '#winningest_coach' do
-      expect(stat_tracker.winningest_coach("20132014")).to eq("Claude Noel")
-    end
-
-    it '#worst_coach' do
-      expect(stat_tracker.worst_coach("20132014")).to eq("Mike Babcock")
-    end
-
-    it '#game_ids_by_season' do
-      expect(stat_tracker.game_ids_by_season("20122013").size).to eq(111)
-    end
-
-    it '#game_teams_by_season' do
-      expect(stat_tracker.game_teams_by_season("20122013").size).to eq(220)
-    end
-
-    it '#coach_stats_by_season' do
-      expect(stat_tracker.coach_stats_by_season("20122013")["John Tortorella"]).to eq([14, 2])
-    end
-
-    it '#team_shots_by_season' do
-      expect(stat_tracker.team_shots_by_season("20122013")["3"]).to eq([19, 100])
-    end
-
-    it '#most_accurate_team' do
-      expect(stat_tracker.most_accurate_team("20132014")).to eq "Utah Royals FC"
-      expect(stat_tracker.most_accurate_team("20142015")).to eq "Toronto FC"
-    end
-
-    it "#least_accurate_team" do
-      expect(stat_tracker.least_accurate_team("20132014")).to eq "Sky Blue FC"
-      expect(stat_tracker.least_accurate_team("20142015")).to eq "North Carolina Courage"
-    end
-
-    it "#tackles_by_season" do
-      expect(stat_tracker.tackles_by_season("20132014")["3"]).to eq 102
-    end
-
-    it "#most_tackles" do
-      expect(stat_tracker.most_tackles("20132014")).to eq "FC Cincinnati"
-      expect(stat_tracker.most_tackles("20142015")).to eq "DC United"
-    end
-
-    it "#fewest_tackles" do
-      expect(stat_tracker.fewest_tackles("20132014")).to eq "Sky Blue FC"
-      expect(stat_tracker.fewest_tackles("20142015")).to eq "North Carolina Courage"
-    end
-  end
 
   context 'Team stats methods' do
     game_path = './spec/fixture_files/test_games.csv'
